scorch zap optimize stored field format for _id field
NOTE: This is a zap version and file format change.

This optimization treats the _id field as a special case, where the
_id field is no longer snappy compressed along with all the other
non-"_id" stored fields.  Instead, the _id field's length and bytes
are written out early and as-is, making the _id value easier to
access.

This can be beneficial to search-related methods, such as
ExternalID(), which only require _id field information.

As part of this, we're also no longer writing out unnecessary data for
the _id field that we persist for other non-"_id" fields, such as: the
numeric fieldID (i.e., uint16(0)), the type (i.e., assumed to be 't'
as a text field), the start offset, the end len, and the
arrayPositions.

Various unit tests were corrected as part of this change in order to
more uniformly treat the _id field as having IndexingOptions as a
stored field and not having DocValues.