scorch optimize conjunctions via push-down to roaring.And()
This change pushes down conjunction information to indexers for
potential TermFieldReader optimizations.  The scorch indexer can
optimize this situation by constructing a more selective roaring
"actual" bitmap for any PostingsIterator's that are part of the
conjunction.

On a bleve-query microbenchmark, dev macbook, with a scorch index of
50K en-wiki docs, using a repeated, single-threaded conjunction query
of several relatively high-frequency terms (query-string of "+text:see
+text:also +text:where +text:http"), perf was ~265 queries/sec before
this change, and ~310 queries/sec after this change.