optimization to recycle DocMatch.FieldTermLocations during search
In this commit, a new field, FieldTermLocations is introduced into the
DocumentMatch struct.  The FieldTermLocations is a slice that is not
JSON serialized and is more ammenable than the existing
DocumentMatch.Locations field for higher performance search
processing, as slices are more friendly than maps for memory reuse.

As the search collector finalizes results, it invokes the newly
introduced DocumentMatch.Complete() method on the final hits, which
converts the DocumentMatch.FieldTermLocations slice into a
DocumentMatch.Locations map.

All scorers and searchers that used to rely on DocumentMatch.Locations
are appropriately updated.  Of note, the phrase searcher also invokes
DocumentMatch.Complete() so that it has the Locations map available
for faster phrase processing.

Microbenchmarks of bleve-query for query-string of 'text:"see also"'
on a 50K en-wiki scorch index went from ~74 q/sec before this change
to ~89 q/sec after this change.