fixes #895 - scorch zap file format updated w/ # bytes of locations
NOTE: This is a zap file format change.

In the case of a composite field (i.e., _all field), the component
fields might have differing configurations for IncludeTermVectors.
This can mean the freq can sometimes be > numberOfLocations, where the
previous code incorrectly assumed freq == numberOfLocations.

By encoding the # of bytes used for locations data, the loops
performing readLocation() will be able to read the correct # of
locations.

And, encoding the # of bytes used for locations data instead of the
count of locations allows for an optimization of no longer reading
unneeded locations data when skipping ahead.

This fixes issue https://github.com/blevesearch/bleve/issues/895
reported by @xeizmendi.