optimize scorch FieldDictFuzzy to check for tooManyClauses earlier
This should not affect functionality, as the fuzzy searcher uses a
disjunction searcher underneath the hood, and the disjunction searcher
later on checks for the tooManyClauses() case.

But, the idea is that the tooManyClauses() check should be handled the
same whether or not the IndexReaderFuzzy interface is implemented by
the index-reader, and if we can bail earlier, then we waste less
resources.

Also, this commit also checks for error more consistently.