optimize scorch to only recycle current TermFieldReaders
When scorch is finished using a TermFieldReader, if the indexSnapshot
that it belongs to is obsolete (not the current indexSnapshot), then
don't recycle the TermFieldReader so that GC can take care of it
sooner.

This can matter in a scenario of ongoing mutations (the current
indexSnapshot keeps on changing) when there are searches (like prefix
& wildcard/regexp) that might involve lots (hundreds or more) of terms
per search request, meaning lots of TermFieldReaders can be in play.

See also: https://issues.couchbase.com/browse/MB-29923