MB-31974 - Streaming search results
Adding optional builder/maker callbacks for the
document match handlers from applications.
These builder method gives back a threadsafe
document match handler callback for bleve which
it will invoke for every document hit during
a search.
The application should make its own copy of
the hit/documentMatch in the callback if it
wish to hold on to data.