optimize scorch postingsIter chunk maintenance on skip ahead
Consider a scenario of a conjunction search with a high-frequency term
search and a low-frequency term search.  The scorch PostingsIterator
for the high-frequency term will have its
nextDocNumAtOrAfter(atOrAfter) method invoked with a non-zero
atOrAfter parameter -- driven by docNum's from the low frequency term
search.

In this case, before this change, the inner loop that moves the
high-frequency "all iterator" forwards performs a division by
chunkFactor on every single "all" docNum, in order to compute the all
iterator's current chunk number, so that it can be checked against the
"acutal iterator"'s current chunk number (in order to maintain the
PostingIterator's freqNorm & loc reader positions).

This integer division was appearing in pprof, so this change is to
replace the division op with a faster integer >= comparison.

On a 200K en-wiki docs perf test, a query-string search for "+http
+schoch" went from ~7650 q/sec throughput before the change to ~8000
q/sec afterwards.