optimize scorch 'unadorned' disjunctions via roaring OR's
Normally, the procesing of disjunctions would need postings iterators
that provide the matching internal-id's along with associated
freq-norm and term-vector information.

A so-called "unadorned" disjunction is a more restricted edge case,
where the application does not need the extra freq-norm information (a
non-scoring disjunction) and does not need term-vector information.

The scorch indexer can optimize these unadorned disjunctions by OR'ing
together the constituent, actual bitmaps, since we do not need to
maintain readers for the freq-norm and term-vector data.  The
resulting OR'ed bitmap is then wrapped by a single term-field-reader
as an alternative to the normal disjunction searcher machinery.

Regarding perf microbenchmarks (bleve-query) on a 200K en-wiki docs
scorch index...

for a high number of high-frequency terms...

- wildcard search on "th*" (~31K hits)...
    before the change, with normal scoring         -  ~4.7 q/sec
    w/ NoScore:true                                -  ~5.1 q/sec
    w/ NoScore:true & unadorned disj. optimization - ~11.6 q/sec

for a low number of high-frequency terms...

- query-string search on "http www com" (~25K hits)...
    before the change, with normal scoring         - ~190 q/sec
    w/ NoScore:true                                - ~260 q/sec
    w/ NoScore:true & unadorned disj. optimization - ~415 q/sec

for a low number of low-frequency terms...

- query-string search on "marty shoch" (207 hits)...
    before the change, with normal scoring         - ~15.0K q/sec
    w/ NoScore:true                                - ~21.3K q/sec
    w/ NoScore:true & unadorned disj. optimization - ~16.1K q/sec

Searching with NoScore:true is faster than normal scoring in all
cases.  However, for low numbers of low-frequency terms, the unadorned
disjunction optimization is slower (~16.1K q/sec) than the classic
disjunction searcher implementation (~21.3K q/sec).  This is likely
due to more memory/garbage creation for the computed OR'ed bitmaps.
This might be addressed in future commits by using heuristics.