optimize unadorned conjunctions (API change)
NOTE: API change -- NewConjunctionSearcher() now returns a
search.Searcher instead of the previous *ConjunctionSearcher

This commit provides an optimization for "unadorned" conjunctions
similar to the previous optimization for unadorned disjunctions.

As part of this, the optimizeCompositeSearcher() func was
renamed/refactored for reuse for both disjunctions and conjunctions.

Regarding perf microbenchmarks (bleve-query) on a 200K document
en-wiki scorch index...

On a query-string search for "+www +com +http" (231 hits)...

  before the change, with normal scoring         -   ~689 q/sec
  w/ NoScore:true                                -  ~2184 q/sec
  w/ NoScore:true & unadorned conj. optimization - ~10148 q/sec