MB-32846 - more aggressively removeOldData() in scorch persister
The scorch persister loop has an optimization to not block/wait if it
sees that the latest root epoch has changed during the current
persistence round, so that the persister can continue immediately to
the top of the persister loop.  But, this "continue OUTER"
optimization would skip a removeOldData() invocation.

This meant that removeOldData() wouldn't be invoked for potentially a
long time in this kind of indexing-heavy scenario.

While indexing 200K wiki docs (indexing only using bleve-blast),
before this change, # of files could grow up to >100 zap segment
files.  After this change, the # of files stayed nearer the
neighborhood of 10~20 files.

See also: https://issues.couchbase.com/browse/MB-32846

Also, cleaned up some whitespace lines seen while trying to diagnose
this issue.