Re-enabling conjunction single clause optimization conditionally
If the conjunction query had a single clause, and this single clause
were to be a match-none searcher, do not apply the optimization
of returning the single searcher itself. This is because if the
conjunction query were a boolean query's must, then when the boolean
searcher is being initialized, it's must searcher would be reset to
nil - there by dropping what could have played a relevant role.