fix issue mapping documents containing *time.Time (#1153)
* fix issue mapping documents containing *time.Time

If documents contained a field of type *time.Time AND they also
had an explicit mapping for the field of a type OTHER THAN text,
the field would not be mapped.

The issue was introduced when we added special support for
handling encoding.TextMarshaler because *time.Time satisfies
this interface.

This change fixes this issue, but there are some more complicated
cases exposed by this, specifically related to mapping the same
field multiple times, and handling cases where you try to mix
multiple types.

Instead, our solution is to further limit the way we support
using ecoding.TextMarshaler.  The new limitation is that we ONLY
use the output of MarshalText() if you have an explicit mapping
AND ALL the index fields are of type text.  All other cases
will ignore TextMarshaler and follow regular indexing rules.

* address code review comments