change sort strategy for integration tests (#1161)
any test which is not explicitly setting a sort order
should instead specify:

"sort": ["-_score", "_id"]

this starts with the natural order of score descending,
but also imposes an additional tiebreaking rule to sort
by the documents identifier.  this comes with some
additional cost, but makes results more stable for
testing purposes.