fixes #1168 - filter duplicate locations in search results
This fix filters out duplicate locations using a destructive in-place
sort on the locations slice, and then a pass to remove duplicates.

This hopefully does not have undue impact on normal case performance,
as the deduplication effort happens only in the "last step" of
finalizeResults() and its DocumentMatch.Complete() invocations.

e.g., the deduplication will be only invoked on the final result hits
(e.g., "size": 10) instead of on all the (potentially many) candidate
document-matches which were examined beforehand.

Also, we only perform the deduplication effort only if we detect that
the locations that are being examined look like they're not in
increasing order.