fixes #1166 - perf optimizations for ReadUvarint()
This commit introduces a specialized MemUvarintReader which is focused
on the "hot inner loop" of scorch/zap's reading of uvarint's -- such
as the postingsList reading of freqNormHasLocs and location data.

The various changes include...

- the separate bytes.Reader vs binary.ReadUvarint() codepaths are
  unified, and ReadByte() is now inlined, along with more local vars.

- SkipUvarint() allows entries to be skipped more efficiently (by
  postingsList.currChunkNext()).

- prevRune maintenance is removed.

- postingsList.readLocation() is always provided a non-nil output
  location.

In microbenchmarks...

  $ go test -benchmem -bench=. ./index/scorch/segment/
  BenchmarkUvarint-8          100000000 16.4  ns/op 0 B/op 0 allocs/op
  BenchmarkMemUvarintReader-8 200000000  6.15 ns/op 0 B/op 0 allocs/op

In bleve-query microbenchmarks on a scorch index with 200K en-wiki
docs, high-frequency term ("http") searches that included locations
had throughput of ~21.5 q/sec before this change, and after this
change went to ~28.8 q/sec.