remove maxTerms optimization for ComputeGeoRange()
Unlike normal disjunctions, a geoboundingbox searcher does not limit
its multi-term-searcher against the DisjunctionMaxClauseCount, so my
recent maxTerms optimization for ComputeGeoRange() in commit 3d774b9
was just wrong.

This commit is a manual revert instead of an automated "git revert",
in that the previous commit 3d774b9 also changed a few things that
we'd still like to keep (e.g., error return value instead of
panic'ing) and there are now more unit tests in this commit from
diagnosing this issue, which compare the outputs of the optimized
ComputeGeoRange() with the previous, non-optimized version.

See also: https://issues.couchbase.com/browse/MB-33614