fix some minor issues with zap doc values (#1202)
- first, we now check the error returned by
loadFieldDocValueReader

- second, in order for that to no break things we changed the
behavior of loadFieldDocValueReader to not return an error
in the case where the field isn't uninverted.  this isn't really
an error, we now just return nil, nil

- third, we return a new error when the condition
    fieldDvLocEnd-fieldDvLocStart > 16
isn't satisfied.  this is a good indication that we're pointing
at the wrong part of the file, or it has been corrupted.

- fourth, the unit test helper method
testMergeWithEmptySegments was updated to check for errors
returned by AnalysisResultsToSegmentBase

- finally, the loadDvReaders metdod was updated to return early
when the number of docs is 0.  this was necessary because there
is a way to build a segment which has an invalid docValueOffset.
this case was exposed by the empty segment building test.  a
proper fix is to would change the file format, so adding this
check was preferred for the moment.