add optional validation hooks to scorch/zap
ValidateDocFields can be used to check each field
seen by zap

ValidateMerge can be used to check the results of
a merge operation

These hooks are intended to allow for application specific
checks inside the main bleve branch, so as to allow maximum
flexibility with minimal disruption.  These hooks are not
intended to be permanent, but that decision can be made
at a later time.