adding UTs for GeoHash
Its been noted a precision/accuracy issue
with respect to the existing geoHash decoder
since it performs rounding of lat/lon values.
This matters when the given hash is having a
smaller precision like 5, eg: "d3hn3".

This test is just a confirmation for the
results with some sample values.