when index is already closed, close returns ErrorIndexClosed
this change make Close() consistent with the other bleve
methods.  if an index is closed, the method should do
nothing, and return ErrorIndexClosed.

previously, the underlying index implementation would
be closed again, and in the case of scorch this can
lead to panics:

ERRO[0004] runtime.plainError close of closed channel
/usr/local/go/src/runtime/panic.go:513 (0x42e079)
	gopanic: reflectcall(nil, unsafe.Pointer(d.fn), deferArgs(d), uint32(d.siz), uint32(d.siz))
/usr/local/go/src/runtime/chan.go:335 (0x4074fe)
	closechan: panic(plainError("close of closed channel"))