ArrayDepthSearcher draft
The intent is to have a new searcher which
helps to search inside slice of objects nested
with in a document.

The implementation takes the approach of having
a FilteringSearcher on top of the conjunction searcher.
The new filtering searcher implementation check whether
all the matched terms in the document are sharing an
array position.
Thinking is that a even single array position is shared
then the document match is a valid one.

Ref: https://forums.couchbase.com/t/fts-inside-array/15476