[BP] MB-37579: [MB-35333 - Potential removeOldData call starve issue]
If the persister is busy doing the catch up with
the ongoing mutations then there is a chance for
the older snapshot clean up call/removeOldData to
wait for a longer time to get invoked. This can
result in unbounded growth in the number of scorch
files.

This change attempt to invoke removeOldData whenever
the number of files grow beyond the thresholds.