Tightening the scorch stale file clean ups
Fix for #1269.
Its observed in an under resourced situation that
the merger was treading ahead of the persister and the
scorch files were piling up. It was reaching upto 70K
with an update heavy workload.

The intention of this fix is to unflip any
ineligible files which are genuinely eligible for
clean up.

Also, adding stats to track this for future.