fix bug where during merge incorrect chunk size is used
this issue would arise when a field has the empty term
and the term has cardinality > 1024.

previously we used some existing logic to detect when we
starte a new term (prevTerm != term), however this logic
fails to differentiate between nil prevTerm (zero value)
and starting the first term (non-nil empty slice).

this distinction was not important for the previous merge
code, because it was just used to finish the old term,
and reset, whereas the new logic is intended to set up
for the coming term, and so must fire at the start as well.

the fix was to handle this case separately and also check
if prevTer is nil