rewrite internal imports
all references to:
github.com/blevesearch/bleve
have been rewritten to:
github.com/blugelabs/bleve

rewrite internal imports

all references to:
github.com/couchbase/vellum
have been rewritten to:
github.com/blugelabs/vellum

switch to blugelabs fork of segment

switch to blugelabs go-porterstemmer

removing manifest, only supporting go modules

refactor interface between scorch and segment impls

The goal is to allow scorch to maintain a map of
supported segment impl types/versions.  And upon
opening an index, check that the required impl is
supported, and if so, use the correct impl.

New interfaces to differentiate between persisted and
unpersisted segments:

Persisted ones have a `Path() string` method
Unpersisted ones have a `Persist(path string) error` method

Many places that used *zap.Segment or *zap.SegmentBase were
migrated to these.

New single unified Merge() method in zap, which takes
[]segment.Segment instead of *Segment or *SegmentBase.
Type assertion is done inside this method (previously done
by scorch layer).  The assumption is that Merge should only
be called with with implementations this package will actually
be able to support.  If you call it with unsupported types,
the code will panic.

ValidateMerge signature changed to work with segment.Segment
instead of the impl structs, expect advanced users can adapt
accordingly.

New OptimizablePostingsIterator interface introduced to
describe optional capabilities of a PostingsIterator for
optimization.

Scorch no longer directly uses zap.<anything> instead, you use
the scorch instances segWrapper in it's place.  This
SegmentWrapper has pointers to required functions in a
upported implementation.

Scorch maintains a map of supported segment types and versions.
This map is populated in an init() function.  Allowing Bleve
packagers to decided which types/versions are going to be
supported by any given release of Bleve.  An open question
is if we should export this, or have some thread-safe way
for application builders to change this on their own.

Scorch also has a notion of a default segment wrapper, this
is simply a pointer to the one which should be used by default
in a new scorch index.

When opening and existing index, at the type the snapshot
is loaded, we first read the snapshot meta-data, which tells us
the segment impl type and version.  We look this up in the map,
and if it is supported, we use that going forward.  If the
type/version used in the snapshot is not supported, we refuse
to load that snapshot.  The existing logic is such that we
will continue to search backwards for a snapshot that we can
open.  This isn't particularly useful right now, but doesn't
seem harmful either.

remove zap from bleve

on this fork, zap is now a top-level project
the zap command-line tool has also been removed
it can be built/used directly from the zap package
and is no longer bundled as a part of the bleve
command (this could still change again in the future)

switch to zap v2

link to zap v1 and v2 at the same time

allow forcing zap version

change DictionaryIterator to not use index.DictEntry

make DocumentFieldTermVisitor local dep for segment

reverting package dependency refactor

simplify segment wrapper interface

refactor into new segment plugin interface

updating segment plugin interface

adding our go.mod and go.sum

additional cleanup

change refs to zap v11 and v12