Support for bounded polygon query…
Attempt to introduce GeoBoundingPolygonQuery
to enable the bounded polygon queries.
Polygon searcher internally finds the bounded
rectangle for the given polygon points and
apply the filter which performs the point-in-polygon
checks.
In this iteration, as simplicity is preferred over
precision(on boundary points), the point-in-polygon check
is based on ray-casting algorithm.