Api to trigger a force merge operation on an index
Introducing a new request merge api to enable customers
conveniently invoke a custom compaction/merge operation
on an existing scorch index. This api would let advanced
bleve users to compact their index all the way down to a
single file segment based index.
The api takes an optional mergePlanOptions argument which
would be applied for the forceful merge operations, while
the original mergePlanOptions property of the index remains
intact after the forceful merge cycle exits.
In the absense of a mergePlanOptions argument, default policy
would be an aggressive one to target a single file segment index
with a maximum segment size of 1B.
In the event of incoming mutations, the forceful merge
operations bail out early than strictly applying the
mergePlanOptions to the index.

Its the caller's responsibility to ensure that there is
enough disk space before invoking this compaction/merge
operation as it needs atleast double the disk space
since its a full compaction.