Offline RollbackPoints/Rollback apis for scorch index
This change comprises of enabling the rollback related scorch APIs
to work only with closed/offline indexes.
Both the API signatures have modified to accept an index path
pointing to the scorch index location to accommodate this change.

The two API approach(RollbackPoints api for fetching the rollback points
and the Rollback api for performing the actual rollback) is retained in
favour of modularity and simplicity of usage.

The Rollback api would make the meta/bolt store with the given
rollback point as the most recent epoch to be loaded.

During the process of restoring the index state to the relevant
rollback point, the new Rollback implementation cleans up the
irrelevant epoch data from the bolt store. This is done to favour
a lean version of persisted epochs in root bolt anytime to further
lead a more efficient disk clean up/utilisation.
Also, this helps to better safe guard against any potential
recursive rollback request as we aren't introducing any newer
data into the system.