Api to trigger a force merge operation on an index
Introducing a new synchronous ForceMerge api to enable
users conveniently invoke a manual merge operation
on an online scorch index. This api would let advanced
bleve users to compact their index all the way down to a
single file segment based index.

The api takes an optional mergePlanOptions argument which
would be applied for the forceful merge operations, while
the original mergePlanOptions property of the index remains
intact after the forceful merge cycle exits.
In the absense of a mergePlanOptions argument, default policy
assumed would be an aggressive one to target a single file
segment index with a maximum segment size of 1B approx.

The api would trigger the merger's work loop only once
during it's invocation and hence it is the caller's
responsibility to actively monitor the expected
file segments count at the root and trigger any
subsequent forceMerge invocations if needed to
meet the expected segment's count at root.

The caller could also pass a cancelCh for cancelling
an ongoing forceMerge operation in case of any
runtime anomalies during the merge operation.

The api also takes an optional parameter to override
the numSnapshotsToKeep for reducing the space wastage
out of segments with duplicate contents.

The caller should also ensure that there is enough disk
space before invoking this compaction/merge operation
as it needs at least double the disk space
whenever there is a full compaction.