optimize TermFacets
The existing TermFacet implementation was O(n) when adding
new TermFacets. All TermFacets were stored in a slice and
the code iterated over the entire slice to determine if
a term already existed.

In MultiSearch, TermFacets for searches across each Index
were merged together, resulting in several iterations over
the same slice.

This change adds a map to `TermFacets` so that existence
checks are done in O(1).