Fix the incorrect behaviour for Composite queries with Score:none
Currently when one of the constituent query is not
optimizable, it just returns the incoming octx without
any modifications.This causes the optimization to proceed
incorrectly as it misses the non optimisable child searcher.

With the fix, when a sub query is not optimizable it
returns a nil context to stop the parent's optimizations.
This would prevent the parent to stop the optimized path
and take the slower path with score.