refactor again to consolidate arguments
I went back to the other point Steve had made about
how many arguments we passed around, and observed that
many of them were never altered.  So, I converted that
into a struct, and made the functions methods on that
struct.

This helped performance get back to match or beat
Steve's version.