fix ineffectual assignment (#1450)
Intention was for calcBudget local var to either use
the CalcBudget provided in the Options, or fall back
to the default mergeplan.CalcBudget method.  However,
after assigning the correct impl to the calcBudget
local var, the code always invoked
mergeplan.CalcBudget anwyay.

This change allows the CalcBudget option to work as
intended.