switch to tagged api versions
note: while doing this it was observed that 'go mod tidy' was
failing with an error.  this relates to the kvstores in the
blevex package which will needed to be updated separately.

this change addresses this by removing the config_* and
upsidedown/benchmark_* files which caused the problem.

this does not remove support for these kvstores, just
the build tag support.