Improve the effectiveness of rollback points in scorch
Scorch originally tries to save the most recent `numSnapshotsToKeep`
snapshots for better rollback-ability. But this is turning out to
be ineffective in capturing any reasonable history of the index.

Proposed approach tries to,
-Save the most recent snapshot/epoch.
-Evenly distribute the remaining `numSnapshotsToKeep-1` points at
 regular intervals in the whole set of persisted list of epochs.

Another potential benefit here is that,

In a heavy traffic system with continuous mutations, the list
of persisted epochs tend to be a lengthy one. This happens as
the removeOldData() clean up happens only when a persister
caught up with the current root.
Evenly distributing rollback points within a lengthy
persisted epoch list is a better bet for rollbackability.