switch default index to scorch with latest zap
one example required updating as it's custom sort contained a
tie (both fields with age 0).  upsidedown always returned the
results in a particular order, which was not guaranteed by scorch.
to address this, the exmaple has been updated to include the
doc id as an additional sort criteria (ensuring the original
order expected is kept by scorch)